Fix Compilation Warnings
git-svn-id: http://localhost:8080/svn/AlgoQuest/trunk@58 de6f410c-f7d5-3442-bb8a-973c24c52ac3

+++ AQQueryResolver/DBLoader/DatabaseLoader.cpp
@@ -60,7 +60,7 @@ int rc;
 
 char name_log [ k_taille_message ];
 //-------------------------------------------------------------------------------
-int cut_in_col (const char * iniFilename, unsigned int num_table, unsigned int num_column) 
+int cut_in_col (const char * iniFilename, size_t num_table, size_t num_column) 
 {
 	// parametres des emplacement des fichiers 
 	char *param_file;
